Fix issue where DocumentOutliner was outputting corrupt utf-8
This was causing the stats file to be emptied because json_encode
was choking on the invalid utf-8. Resulting json error message:
Malformed UTF-8 characters, possibly incorrectly encoded

The issue occurred in DocumentOutliner::collapseWhitespace in the
preg_replace function. Apparently preg_replace used with \s does not
play well with multi-byte strings.